# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "kasrchem System",
        "description": "kasrchem API Documentation",
        "contact": {
            "email": "admin@kasrchem.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend.kasrchem.murabba.dev/api/v2",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/photos_galleries": {
            "get": {
                "tags": [
                    "Multimedia"
                ],
                "summary": "Get list of Photos Galleries",
                "operationId": "Get list of Photos Galleries",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/photos_galleries/{id}": {
            "get": {
                "tags": [
                    "Multimedia"
                ],
                "summary": "Show Gallery Data",
                "operationId": "show_gallery",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Gallery ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/latest_galleries": {
            "get": {
                "tags": [
                    "Multimedia"
                ],
                "summary": "Get list of Latest gallery limit 15",
                "operationId": "Get list of Latest gallery limit 15",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/latest_releases": {
            "get": {
                "tags": [
                    "Multimedia"
                ],
                "summary": "Get list of Latest releases limit 15",
                "operationId": "Get list of Latest releases limit 15",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/latest_posts": {
            "get": {
                "tags": [
                    "Multimedia"
                ],
                "summary": "Get list of Latest posts limit 15",
                "operationId": "Get list of Latest posts limit 15",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/latest_videos": {
            "get": {
                "tags": [
                    "Multimedia"
                ],
                "summary": "Get list of Latest videos limit 15",
                "operationId": "Get list of Latest videos limit 15",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "Pass Bearer Token",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}